chore: add code comments
