+++ config_file.conf
@@ -1,8 +1,8 @@
 server {
-	listen 127.0.0.1:8080;
+	listen 0.0.0.0:8080;
 	server_name server_1;
 
-	root /home/quahxuanyu/42KL/42-webserv;
+	root /Users/joophang/Desktop/42/webserv;
 
 	error_page 400 /error_page/400.html;
 	error_page 401 /error_page/401.html;
@@ -16,23 +16,23 @@ server {
 	error_page 504 /error_page/504.html;
 	error_page 505 /error_page/505.html;
 
-	#client_body_size_buf 100000000;
+	# client_body_size_buf 100000000;
 
 	#test root/index
 	location / {
-		root /home/quahxuanyu/42KL/42-webserv;
+		root /Users/joophang/Desktop/42/webserv;
 		index html/main.html;
 		methods GET POST;
 	}
 
 	#test alias
 	location /test1 {
-		alias /home/quahxuanyu/42KL/42-webserv;
+		alias /Users/joophang/Desktop/42/webserv;
 		methods GET POST;
 	}
 
 	location /upload {
-		alias /home/quahxuanyu/42KL/42-webserv;
+		alias /Users/joophang/Desktop/42/webserv;
 		index html/upload.html;
 		methods GET POST;
 	}
@@ -43,24 +43,24 @@ server {
 	}
 
 	location /autoindex {
-		alias /home/quahxuanyu/42KL/42-webserv/;
+		alias /Users/joophang/Desktop/42/webserv/;
 		autoindex on;
 		methods GET;
 	}
 
 	location /list {
-		alias /home/quahxuanyu/42KL/42-webserv/;
+		alias /Users/joophang/Desktop/42/webserv/;
 		index html/list-files.html;
 		methods GET;
 	}
 
 	location /cgi-bin {
-		alias /home/quahxuanyu/42KL/42-webserv/cgi-bin;
+		alias /Users/joophang/Desktop/42/webserv/cgi-bin;
 		methods GET POST DELETE;
 	}
 
 	location /login {
-		alias /home/quahxuanyu/42KL/42-webserv;
+		alias /Users/joophang/Desktop/42/webserv;
 		index html/login.html;
 		methods GET POST;
 	}
@@ -74,7 +74,7 @@ server {
 	client_body_size_buf 100000000;
 
 	location / {
-		root /home/quahxuanyu/42KL/42-webserv;
+		root /Users/joophang/Desktop/42/webserv;
 		index html/sample2.html;
 		autoindex on;
 		methods GET POST ;
@@ -85,21 +85,21 @@ server {
 	listen 127.0.0.1:2323;
 	#server_name server_1;
 
-	root /home/quahxuanyu/42KL/42-webserv;
+	root /Users/joophang/Desktop/42/webserv;
 
 	error_page 404 /error_page/404.html;
 	client_body_size_buf 100000000;
 
 	location / {
 
-		root /home/quahxuanyu/42KL/42-webserv;
+		root /Users/joophang/Desktop/42/webserv;
 		index html/login.html;
 		autoindex on;
 		methods POST GET;
 	}
 	
 	location /login {
-		alias /home/quahxuanyu/42KL/42-webserv;
+		alias /Users/joophang/Desktop/42/webserv;
 		index html/login.html;
 		methods GET POST;
 	}
@@ -108,13 +108,13 @@ server {
 server {
 	listen 127.0.0.1:2626;
 
-	root /home/quahxuanyu/42KL/42-webserv;
+	root /Users/joophang/Desktop/42/webserv;
 
 	error_page 404 /error_page/404.html;
 	client_body_size_buf 100000000;
 
 	location / {
-		root /home/quahxuanyu/42KL/42-webserv;
+		root /Users/joophang/Desktop/42/webserv;
 		index html/sample2.html;
 		autoindex off;
 		methods GET POST ;
