+++ test.cpp
@@ -1,0 +1,7 @@
+#include <iostream>
+
+int main()
+{
+	bool boo;
+	std::cout << boo << std::endl;
+}
